fix: rename add command
