+++ cli/src/commands/add.fsm.ts
